Project: Add KASLR Warning
This commit adds a warning for users with KASLR disabled if it was disabled
unintentionally, it warns of any potential risks of running the device with it
disabled and how to re-enable it.

Also while at it, move the string padding function to exynos_font and add a
missing font height define to the lk3rd display header.

Signed-off-by: Umer Uddin <umer.uddin@mentallysanemainliners.org>